fix: sync database schema + fix dev server startup
- Create 36 APP-ONLY tables in Supabase
- Add 102 new columns across 20+ tables via ALTER TABLE
- Generate drizzle migration (0000_strong_jean_grey.sql)
- Fix z.string().uuid().int() invalid Zod chains
- Dev server boots successfully

+++ server/estimate-router.ts
@@ -753,7 +753,7 @@ export const estimateRouter = router({
   listPartialDrafts: protectedProcedure
     .input(
       z.object({
-        scopeDraftId: z.string().uuid().int().optional(),
+        scopeDraftId: z.string().uuid().optional(),
         status: z.enum(["pending", "retrying", "recovered", "abandoned"]).optional(),
         limit: z.number().int().min(1).max(100).optional(),
         offset: z.number().int().min(0).optional(),
